+++ src/main/java/com/github/lipiridi/spotless/applier/action/ReformatFileAction.java
@@ -7,7 +7,6 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
-import com.intellij.psi.PsiFile;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
 
@@ -22,15 +21,14 @@ public class ReformatFileAction extends AnAction {
 
         Optional.ofNullable(event.getDataContext().getData(CommonDataKeys.EDITOR))
                 .map(Editor::getDocument)
-                .ifPresent(document -> {
-                    PsiFile psiFile = event.getDataContext().getData(CommonDataKeys.PSI_FILE);
-                    new ReformatProcessor(project, document, psiFile).run();
-                });
+                .ifPresent(
+                        document -> Optional.ofNullable(event.getDataContext().getData(CommonDataKeys.PSI_FILE))
+                                .ifPresent(psiFile -> new ReformatProcessor(project, document, psiFile).run()));
     }
 
     @Override
     public @NotNull ActionUpdateThread getActionUpdateThread() {
-        return super.getActionUpdateThread();
+        return ActionUpdateThread.BGT;
     }
 
     @Override