+++ src/main/java/com/github/lipiridi/spotless/applier/ui/settings/SpotlessApplierSettingsConfigurable.java
@@ -1,5 +1,6 @@
-package com.github.lipiridi.spotless.applier.ui;
+package com.github.lipiridi.spotless.applier.ui.settings;
 
+import com.github.lipiridi.spotless.applier.ui.SpotlessApplierSettingsComponent;
 import com.intellij.openapi.options.Configurable;
 import javax.swing.*;
 import org.jetbrains.annotations.Nls;
@@ -28,15 +29,17 @@ public class SpotlessApplierSettingsConfigurable implements Configurable {
     @Override
     public boolean isModified() {
         SpotlessApplierSettingsState settings = SpotlessApplierSettingsState.getInstance();
-        return (spotlessApplierSettingsComponent.isOptimizeImportsBeforeApplying()
-                        != settings.optimizeImportsBeforeApplying)
-                || (spotlessApplierSettingsComponent.isProhibitAsteriskImports()
-                        != settings.prohibitImportsWithAsterisk);
+
+        return spotlessApplierSettingsComponent.isAllowGradleCache() != settings.allowGradleCache
+                ^ spotlessApplierSettingsComponent.isOptimizeImportsBeforeApplying()
+                        != settings.optimizeImportsBeforeApplying
+                ^ spotlessApplierSettingsComponent.isProhibitAsteriskImports() != settings.prohibitImportsWithAsterisk;
     }
 
     @Override
     public void apply() {
         SpotlessApplierSettingsState settings = SpotlessApplierSettingsState.getInstance();
+        settings.allowGradleCache = spotlessApplierSettingsComponent.isAllowGradleCache();
         settings.optimizeImportsBeforeApplying = spotlessApplierSettingsComponent.isOptimizeImportsBeforeApplying();
         settings.prohibitImportsWithAsterisk = spotlessApplierSettingsComponent.isProhibitAsteriskImports();
     }
@@ -44,6 +47,7 @@ public class SpotlessApplierSettingsConfigurable implements Configurable {
     @Override
     public void reset() {
         SpotlessApplierSettingsState settings = SpotlessApplierSettingsState.getInstance();
+        spotlessApplierSettingsComponent.setAllowGradleCache(settings.allowGradleCache);
         spotlessApplierSettingsComponent.setOptimizeImportsBeforeApplying(settings.optimizeImportsBeforeApplying);
         spotlessApplierSettingsComponent.setProhibitAsteriskImports(settings.prohibitImportsWithAsterisk);
     }