+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## [Unreleased]
 
+### Added
+- Support for multi-module projects
+- Option that allow Gradle's configuration cache for 'apply' task
+
+### Changed
+- Rename task "Reformat All Files With Spotless" to "Reformat Project With Spotless"
+
 ## [1.0.1] - 2023-08-11
 
 ### Added